+++ R/model-normal.R
@@ -1,6 +1,8 @@
 #' Normal Regression for Continuous Dependent Variables
 #'
-#' Vignette: \url{http://docs.zeligproject.org/articles/zelig_normal.html}
+#' @param formula a model fitting formula
+#'
+#' @seealso Vignette: \url{http://docs.zeligproject.org/articles/zelig_normal.html}
 #' @import methods
 #' @export Zelig-normal
 #' @exportClass Zelig-normal
@@ -32,7 +34,7 @@ znormal$methods(
   param = function(z.out, method="mvn") {
     degrees.freedom <- z.out$df.residual
     sig2 <- base::summary(z.out)$dispersion # not to call class summary method
-    simalpha <- sqrt(degrees.freedom * sig2 
+    simalpha <- sqrt(degrees.freedom * sig2
                      / rchisq(.self$num, degrees.freedom))
 
     if(identical(method,"mvn")){